+++ CHANGELOG.md
@@ -26,6 +26,9 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 * Make `trim_grids()` a public method of `OptimizationGrid` and update it to
   remove rows of NaNs.
 
+* Raise a `ValueError` if any media channel have all zeros or all `NaN`
+  impressions.
+
 ## [1.1.7] - 2025-07-16
 
 * Fix `rhat_summary()` to work with a vector sigma dim.